Move injected document repository to base class
