+++ app/Console/GeocodeEventCommand.php
@@ -6,27 +6,11 @@ use Broadway\CommandHandling\CommandBusInterface;
 use CultuurNet\UDB3\Address\Address;
 use CultuurNet\UDB3\Event\Commands\UpdateGeoCoordinatesFromAddress;
 use CultuurNet\UDB3\Event\ReadModel\DocumentGoneException;
-use CultuurNet\UDB3\ReadModel\DocumentRepository;
-use CultuurNet\UDB3\Search\ResultsGeneratorInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class GeocodeEventCommand extends AbstractGeocodeCommand
 {
-    /**
-     * @var DocumentRepository
-     */
-    private $documentRepository;
-
-    public function __construct(
-        CommandBusInterface $commandBus,
-        ResultsGeneratorInterface $searchResultsGenerator,
-        DocumentRepository $documentRepository
-    ) {
-        parent::__construct($commandBus, $searchResultsGenerator);
-        $this->documentRepository = $documentRepository;
-    }
-
     /**
      * @inheritdoc
      */
@@ -56,7 +40,7 @@ class GeocodeEventCommand extends AbstractGeocodeCommand
     {
 
         try {
-            $document = $this->documentRepository->get($eventId);
+            $document = $this->getDocumentRepository()->get($eventId);
         } catch (DocumentGoneException $e) {
             $document = null;
         }
